Check_POST_REQ_NOT_WORKING

+++ app/src/main/java/com/example/saneforce_task/Product.kt
@@ -1,14 +1,12 @@
 package com.example.saneforce_task
 
 
-import com.google.gson.annotations.SerializedName
-
 data class Product(
         val product_code: String,
         val product_name: String,
         var Product_Qty: Int,
-        val Rate: Int,
-        var Product_Amount: Int
+        val Rate: Double,
+        var Product_Amount: Double
     )
 
 data class ProductRequest(
